capstone/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
supabase==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.0
google-generativeai==0.3.0
pytesseract==0.3.10
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
pgvector==0.2.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
SQLAlchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0
